Do not load plugin on PHP < 7.1
