+++ src/ComposerPlugin.php
@@ -2,33 +2,8 @@
 
 namespace Cs278\ComposerAudit;
 
-use Composer\Composer;
-use Composer\IO\IOInterface;
-use Composer\Plugin\PluginInterface;
-use Composer\Plugin\Capable;
-use Composer\Plugin\Capability\CommandProvider as CommandProviderCapability;
-
-final class ComposerPlugin implements PluginInterface, Capable
-{
-    public function activate(Composer $composer, IOInterface $io)
-    {
-
-    }
-
-    public function deactivate(Composer $composer, IOInterface $io)
-    {
-
-    }
-
-    public function uninstall(Composer $composer, IOInterface $io)
-    {
-
-    }
-
-    public function getCapabilities()
-    {
-        return array(
-            CommandProviderCapability::class => CommandProvider::class,
-        );
-    }
+if (\PHP_VERSION_ID >= 70100) {
+    require __DIR__.'/ComposerPlugin.real.php';
+} else {
+    require __DIR__.'/ComposerPlugin.fake.php';
 }
